Update: change excute points
