+++ tsconfig.json
@@ -4,21 +4,10 @@
     "target": "es5",
     "module": "commonjs",
     "sourceMap": false,
-    "types": [
-      "@types/chrome",
-      "node"
-    ],
-    "typeRoots": [
-      "node_modules/@types"
-    ],
+    "types": ["@types/chrome", "node"],
+    "typeRoots": ["node_modules/@types"],
     "moduleResolution": "node",
-    "lib": [
-      "es2017",
-      "dom"
-    ]
+    "lib": ["es2017", "dom"]
   },
-  "include": [
-    "src/background",
-    "src/contentscript"
-  ]
+  "include": ["./src/*"]
 }
